third_party/WebKit/LayoutTests/external/wpt/html/semantics/forms/resetting-a-form/reset-form-2-expected.txt

This is a testharness.js-based test.
FAIL integration test on reset for a created-from-script form assert_equals: option should reset dirtyness to false expected false but got true
Harness: the test ran to completion.
